change the number type of linearize again

+++ num_gradient/num_gradient_descent_master/t_ngd_cifar10.py
@@ -260,7 +260,7 @@ def create_f(h, w, target):
 #     return h, w, img_array
 
 def linearize_pixels(img):
-    x = np.copy(np.asarray(img, dtype='uint8'))  # 保持为 uint8 类型
+    x = np.copy(np.asarray(img))  # 保持为 uint8 类型
     h, w, c = x.shape
     img_array = x.reshape(h * w * c).astype('uint8')  # 确保仍然是 uint8 类型
     return h, w, img_array
